+++ src/super/i-block/modules/render-info/CHANGELOG.md
@@ -8,3 +8,9 @@ Changelog
 > - :memo:       [Documentation]
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
+
+## XXX (XXX)
+
+#### :rocket: New Feature
+
+* Initial release
