+++ src/super/i-block/modules/render-info/interface.ts
@@ -43,5 +43,5 @@ export type RenderStrategy = (
 	/**
 	 * Data for rendering
 	 */
-	debugDataStore: Dictionary
+	data: Dictionary
 ) => Promise<void>;
